python/requirements.txt
Flask
Flask_restful
flask-mysql-connector